Deprecate `str::from_utf16_lossy`
Use `String::from_utf16_lossy` instead.

[breaking-change]
